3d model load 할 수 있는 기능 추가 obj.파일 로드하는 소스 추가해서 구현 다른 파일들의 코드 정리 출처: https://github.com/tinyobjloader/tinyobjloader

+++ app/source/engine/VKinputEvent.cpp
@@ -33,6 +33,7 @@ namespace vkengine
 
             }
         }
+        
         void mouseButtonCallback(GLFWwindow* window, int button, int action, int mods)
         {
             VulkanEngine* app = reinterpret_cast<VulkanEngine*>(glfwGetWindowUserPointer(window));
@@ -51,6 +52,7 @@ namespace vkengine
             }
             //else if (button == )
         }
+        
         void cursorPositionCallback(GLFWwindow* window, double xpos, double ypos)
         {
             glfwGetCursorPos(window, &xpos, &ypos);
@@ -62,7 +64,7 @@ namespace vkengine
 
             if (app->getCameraMoveStyle())
             {
-                app->getCamera()->RotateScreenStandard(xpos, ypos, windowWidth, windowHeight);
+                app->getCamera()->RotateScreenStandard(static_cast<float>(xpos), static_cast<float>(ypos), windowWidth, windowHeight);
             }
             else
             {
@@ -88,6 +90,7 @@ namespace vkengine
             }
 
         }
+        
         void scroll_callback(GLFWwindow* window, double xoffset, double yoffset)
         {
             VulkanEngine* app = reinterpret_cast<VulkanEngine*>(glfwGetWindowUserPointer(window));
